refactor: refactored TopCastScreen.
